pyproject.toml
[tool.poetry]
name = "vault-k8s-mapper"
version = "0.2.4"
description = ""
authors = ["Your Name <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
hvac = "^0.9.6"
kubernetes = "^10.0"
pyyaml = "^5.3"
python-dotenv = "^0.10.5"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0", allow-prereleases = true}
pylint = "^2.4.4"
pytest = "^5.3.4"
coverage = "^5.0.3"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
